Removed secrets from docker-compose
