+++ apps/atr-one/app/drivers/multiprocess_driver.rb
@@ -1,3 +1,5 @@
+require 'tracing'
+
 class MultiprocessDriver
   include AtrOne::Deps[:logger]
   include Tracing
