+++ apps/atr-one/app/event_brokers/pubnub_event_broker.rb
@@ -1,0 +1,26 @@
+require 'tracing'
+
+class PubnubEventBroker
+  include AtrOne::Deps[:logger]
+  include Tracing
+
+  include AtrOne::Deps[:pubnub]
+
+  def publish(event)
+    pubnub.publish(
+      channel: 'at-the-races',
+      message: { id: event.id, payload: event.payload }
+    ) do |envelope|
+      trace "Published. Response: #{envelope.response}. Status: #{envelope.status}"
+    end
+  end
+
+  alias on_update_allocated publish
+  alias on_lock_acquired publish
+  alias on_lock_released publish
+  alias on_race_delay_starting publish
+  alias on_race_delay_ended publish
+  alias on_trace publish
+  alias on_driver_start publish
+  alias on_driver_stop publish
+end
