+++ apps/atr-one/app/processors/database_locking_processor.rb
@@ -1,3 +1,5 @@
+require 'tracing'
+
 class DatabaseLockingProcessor
   include AtrOne::Deps[:logger]
   include Tracing
