+++ apps/atr-one/config/system/boot/pubnub.rb
@@ -1,25 +1,16 @@
 # Be sure to name this file after the component. Otherwise you see
 # Bootable component :xxx was already registered (Dry::System::DuplicatedComponentKeyError) errors.
 
+# The YAML file /run/secrets/pubnub should look something like this:
+# 
+# ---
+# subscribe_key: <subscribe key from pubnub keyset>
+# publish_key: <publish key from pubnub keyset>
+# user_id: <pubnub user id>
+
 AtrOne::Container.boot(:pubnub) do |container|
-  init do    
-    Pubnub.new(
-      subscribe_key: ENV['PUBNUB_SUBSCRIBE_KEY'],
-      publish_key: ENV['PUBNUB_PUBLISH_KEY'],
-      secret_key: ENV['PUBNUB_SECRET_KEY'],
-      user_id: ENV['PUBNUB_USER_ID']
-    ).tap { |pubnub| register(:pubnub, pubnub) }
-  end
-
-  start do
-    # container[:pubnub].tap do |broker|
-    #   container[:message_bus].subscribe(broker)
-    # end
-  end
-
-  stop do
-    # container[:pubnub].tap do |broker|
-    #   container[:message_bus].unsubscribe(broker)
-    # end
+  init do
+    secrets = YAML.load_file('/run/secrets/pubnub')
+    register(:pubnub, Pubnub.new(secrets.symbolize_keys))
   end
 end
