+++ docker-compose.yml
@@ -3,6 +3,8 @@ version: '3.6'
 services:
   one:
     image: willnotwish/at-the-races-dev:2.5.8
+    secrets:
+      - pubnub
     environment:
       RAILS_LOG_TO_STDOUT: 1
       DATABASE_HOST: mysql55
@@ -11,11 +13,6 @@ services:
       DATABASE_PASSWORD: root
       RAILS_MAX_THREADS: 10
       REDIS_URL: redis://redis:6379/11
-      ABLY_API_KEY: JGNyew.ZUEUWQ:uLlr1i7zv8hFSPH7UuOToBQcp-kK8NCwCuylYWf38ts
-      PUBNUB_SUBSCRIBE_KEY: sub-c-c8ab435c-983f-4236-80e2-531f6c888eaa
-      PUBNUB_PUBLISH_KEY: pub-c-7ee77e5d-96fc-4c7f-a452-85c71713100c
-      PUBNUB_SECRET_KEY: sec-c-NmI0ZTA5NmEtODU1Yy00MjI5LTllN2EtNjAxZmFhNmU5ZjRi
-      PUBNUB_USER_ID: at-the-races-user
     stdin_open: true
     tty: true
     user: nick
@@ -47,6 +44,8 @@ services:
 
   sidekiq:
     image: willnotwish/at-the-races-dev:2.5.8
+    secrets:
+      - pubnub
     environment:
       RAILS_LOG_TO_STDOUT: 1
       DATABASE_HOST: mysql55
@@ -55,11 +54,6 @@ services:
       DATABASE_PASSWORD: root
       RAILS_MAX_THREADS: 10
       REDIS_URL: redis://redis:6379/11
-      ABLY_API_KEY: JGNyew.ZUEUWQ:uLlr1i7zv8hFSPH7UuOToBQcp-kK8NCwCuylYWf38ts
-      PUBNUB_SUBSCRIBE_KEY: sub-c-c8ab435c-983f-4236-80e2-531f6c888eaa
-      PUBNUB_PUBLISH_KEY: pub-c-7ee77e5d-96fc-4c7f-a452-85c71713100c
-      PUBNUB_SECRET_KEY: sec-c-NmI0ZTA5NmEtODU1Yy00MjI5LTllN2EtNjAxZmFhNmU5ZjRi
-      PUBNUB_USER_ID: at-the-races-user
 
     stdin_open: true
     tty: true
@@ -84,6 +78,10 @@ services:
   #   image: apache/kafka:3.9.0
   #   ports:
   #     - 3317:3306
+
+secrets:
+  pubnub:
+    file: ./apps/atr-one/config/secrets/pubnub.yml
 
 volumes:
   atr-one-bundle:
